+++ lib/wrapper_plugins/janus_audio_bridge_plugin.dart
@@ -213,7 +213,7 @@ class JanusAudioBridgePlugin extends JanusPlugin {
       "group": group
     }..removeWhere((key, value) => value == null);
     if (offer == null) {
-      offer = await this.createOffer(videoSend: false, videoRecv: false, audioSend: true, audioRecv: true);
+      offer = await this.createOffer(videoRecv: false, audioRecv: true);
     }
     JanusEvent response = JanusEvent.fromJson(await this.send(data: payload, jsep: offer));
     JanusError.throwErrorFromEvent(response);