+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.3.0
+- breaking changes in createOffer and createAnswer (removed dead code prepareTransReceiver)
+- fixed audio and video mute events not working due to #120 raised by @liemfs 
+
 ## 2.2.15
 - added support for simulcasting in `initMediaDevice`
 - flutter webrtc 0.9.34